+++ src/types/markdown-it-table-of-contents.d.ts
@@ -1,0 +1,5 @@
+declare module 'markdown-it-table-of-contents' {
+    import MarkdownIt from 'markdown-it';
+    const plugin: MarkdownIt.PluginSimple;
+    export default plugin;
+}